Uploaded flight director code
As it stands, the flight director currently has all of its core logic. This includes:
 -Go to a preset coordinate, and drop a pod
 -Return home for battery changing
 -When told to do so, leave and retrieve the pod

However, the flight director is missing most of its major functions (LoRa to find the pod, close-range alignment of the drone over the pod, etc)

The flight director also has two test functions that can be used to ensure functionality of the vehicle.